pyproject.toml
[tool.poetry]
name = "py-soundboard"
version = "0.1.0"
description = ""
authors = ["d42 <d42@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask-restful = "^0.3.7"
prometheus-client = "^0.7.1"
pyyaml = "^5.1"
requests = "^2.22"
python-dateutil = "^2.8"
pysdl2 = "^0.9.6"
paho-mqtt = "^1.5"
evdev = "^1.2"
marshmallow = "^3.2"
blinker = "^1.4"
argumentize = {git = "https://github.com/d42/argumentize"}

[tool.poetry.dev-dependencies]
pytest-httpbin = "^1.0"
pudb = "^2019.1"
pytest-pudb = "^0.7.0"
pyinstaller = "^3.5"
staticx = "^0.7.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
